requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1
django-activity-stream==1.4.2
django-allauth==0.53.1
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-htmx==1.14.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
jsonfield==3.1.0
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4
sqlparse==0.4.3
swapper==1.3.0
urllib3==1.26.15
